Put all hashing algorithms behind Cabal flags
This allows a developer depending on `password` to toggle on and off
support for different password-hashing algorithms at the package level.

The motivator for this is that `cryptonite`'s Argon2 implementation is
broken on M1 Macs (or any `aarch64-darwin` machine), and in the Nix
package set, `cryptonite` has been patched so that any calls to the
Argon2 implementation will cause a compiler error telling the developer
that it is broken. However, as a user of `password`, I don't use Argon2,
and I still want to be able to compile `password` by simply ignoring
Argon2 as a possible hashing algorithm. That's what these flags allow me
to do.