Expose Data.Password.Internal source code to tests
The tests previously relied on the Bcrypt module being exposed to access
certain re-exports from the Internal module. Now that Bcrypt is no
longer guaranteed to be included in the library (i.e. if the brcrypt
cabal flag is disabled), we can't rely on it to be able to run the
tests.

Instead, we add the `src/` folder to the test suite, so that the tests
can directly access the Internal module without needing to expose it to
public users of the library.

This change generates some spurious warnings about "undeclared modules"
when building the tests, but these can safely be ignored. Those modules
are accessed through the `password` library, not via the source code.