Only test scrypt compatibility on x86_64
`password` contains a test that the Scrypt support in `password` is
compatible with the actual `scrypt` Haskell package.

However, the `scrypt` Haskell package only runs on x86_64 (with the SSE2
instruction set).

This commit updates the Scrypt compatibility tests in `password` to only
be included if we're building on x86_64.

See
https://github.com/cdepillabout/password/pull/85#pullrequestreview-2423625976
for more information.